Now to delete,update and retrieve a post the user has to be the same who created that post before

+++ posts/views.py
@@ -3,6 +3,7 @@
 from typing import Iterable, Any
 
 from django.shortcuts import get_object_or_404
+from django.contrib.auth.models import AbstractBaseUser, AnonymousUser
 
 from rest_framework.views import APIView
 from rest_framework.response import Response
@@ -39,9 +40,17 @@ class PostListView(APIView):
 class PostDetailView(APIView):
     """View for each post"""
 
-    def get(self, _request: Request, post_id: int) -> Response:
+    def get(self, request: Request, post_id: int) -> Response:
         """Get one post"""
-        post: Post = Post.objects.get(pk=post_id)
+        post: Post = get_object_or_404(Post, pk=post_id)
+        user: AbstractBaseUser | AnonymousUser = request.user
+
+        if post.author.pk != user.pk:
+            permission_error: dict[str, Any] = {
+                "error": "You do not have enough permissions to retrieve this post "
+            }
+            return Response(data=permission_error, status=status.HTTP_401_UNAUTHORIZED)
+
         serializer: PostSerializer = PostSerializer(instance=post)
         return Response(data=serializer.data, status=status.HTTP_200_OK)
 
@@ -49,15 +58,31 @@ class PostDetailView(APIView):
         """Update a post"""
         data: dict[str, Any] = request.data
         post: Post = get_object_or_404(Post, pk=post_id)
+        user: AbstractBaseUser | AnonymousUser = request.user
+
+        if post.author.pk != user.pk:
+            permission_error: dict[str | Any] = {
+                "error": "You do not have enough permissions to update this posts"
+            }
+            return Response(data=permission_error, status=status.HTTP_401_UNAUTHORIZED)
+
         serializer: PostSerializer = PostSerializer(instance=post, data=data)
         if serializer.is_valid():
             serializer.save()
             return Response(data=serializer.data, status=status.HTTP_200_OK)
         return Response(data=serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
-    def delete(self, _request: Request, post_id: int) -> Response:
+    def delete(self, request: Request, post_id: int) -> Response:
         """Delete a post"""
         post: Post = get_object_or_404(Post, pk=post_id)
+        user: AbstractBaseUser | AnonymousUser = request.user
+
+        if post.author.pk != user.pk:
+            permission_error: dict[str | Any] = {
+                "error": "You do not have enough permissions to delete this posts"
+            }
+            return Response(data=permission_error, status=status.HTTP_401_UNAUTHORIZED)
+
         post.delete()
         response: dict[str, str] = {"Message": "Deleted"}
         return Response(data=response, status=status.HTTP_200_OK)
